Major refactoring removing SFArrayBase, SFList, and SFString classes and using vector, map, and string classes instead. This is a breaking change.

+++ src/tools/getBloom/options.h
@@ -32,4 +32,4 @@ public:
 };
 
 //-----------------------------------------------------------------------------
-extern uint32_t foundFile(const string_q& path, void *data);
+extern size_t foundFile(const string_q& path, void *data);
